requirements.txt
art==5.5
astroid==2.11.2
click==8.1.2
dill==0.3.4
flake8==4.0.1
Flask==2.1.1
Flask-Cors==3.0.10
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
Pillow==9.1.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.13.5
six==1.16.0
tomli==2.0.1
Werkzeug==2.1.1
wrapt==1.14.0
art==5.5
astroid==2.11.2
click==8.1.2
dill==0.3.4
flake8==4.0.1
Flask==2.1.1
Flask-Cors==3.0.10
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
Pillow==9.1.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.13.5
six==1.16.0
tomli==2.0.1
Werkzeug==2.1.1
wrapt==1.14.0
art==5.5
astroid==2.11.2
click==8.1.2
dill==0.3.4
flake8==4.0.1
Flask==2.1.1
Flask-Cors==3.0.10
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
Pillow==9.1.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.13.5
six==1.16.0
tomli==2.0.1
Werkzeug==2.1.1
wrapt==1.14.0
art==5.5
astroid==2.11.2
click==8.1.2
dill==0.3.4
flake8==4.0.1
Flask==2.1.1
Flask-Cors==3.0.10
Flask-WTF==1.0.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
Pillow==9.1.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.13.5
python-decouple==3.6
python-dotenv==0.20.0
six==1.16.0
tomli==2.0.1
Werkzeug==2.1.1
wrapt==1.14.0
WTForms==3.0.1
art==5.5
astroid==2.11.2
click==8.1.2
dill==0.3.4
flake8==4.0.1
Flask==2.1.1
Flask-Assets==2.0
Flask-Cors==3.0.10
Flask-WTF==1.0.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
numpy==1.22.3
Pillow==9.1.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.13.5
pyScss==1.4.0
python-decouple==3.6
python-dotenv==0.20.0
six==1.16.0
tomli==2.0.1
webassets==2.0
Werkzeug==2.1.1
wrapt==1.14.0
WTForms==3.0.1
